renamed nci60 data
